Finalized notifications, changeLogs, user view histories, finalized image upload, implemented index's image-links for software details, finalized similar software, finalized seach features in user software index view, updated smtp client, implemented notification for user views

+++ Vendor Application Inventory Platform/Areas/Admin/Data/Services/CompanyServices.cs
@@ -134,7 +134,7 @@ public class CompanyServices : ICompanyServices
             var contact = new ContactNumber()
             {
                 CityID = city.CityID,
-                Number = int.Parse(contactNum)
+                Number = long.Parse(contactNum)
             };
 
             _db.ContactNumbers.Add(contact);
